Don't pass charset from headers on to fields.
If you do then all the headers after the Content-Type: header will
inherit any encoding set on that line.

At the moment Mail::Field.new defaults to UTF-8 which is arguably the
correct value (according to RFC 5335).

We could attempt to keep headers as BINARY if we're not mutating them,
however this would require annoyingly large changes; and the practical
benefit is not clear as the mail gem seems to silently work with invalid
UTF-8 in field values (it uses rfc-2047 encoding with "ASCII-8BIT",
and while I doubt that that is a valid thing to do, it probably doesn't
matter for us).